UploadVideoLogs/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "experiment-drive-uploader"
version = "0.1.0"
description = "Upload experiment log files to Google Drive folders"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-api-python-client",
  "google-auth-httplib2",
  "google-auth-oauthlib",
  "python-dotenv"
]

[tool.setuptools.packages.find]
include = ["uploader*"]